Change version codes from 32-bit to 64-bit integers
This change coincides with accrescent/directory-api@2f74652 and
accrescent/server-events@d08f59e, changing version codes from a 32-bit
representation to a 64-bit representation to accomodate the full range
of version codes supported by the Android platform.

+++ directory/src/main/kotlin/app/accrescent/server/directory/AppEditPublicationRequestedProcessor.kt
@@ -83,7 +83,7 @@ class AppEditPublicationRequestedProcessor(
                             id = releaseChannelId,
                             appId = event.edit.app.appId,
                             name = it.releaseChannel.canonicalForm(),
-                            versionCode = it.packageMetadata.versionCode.toUInt(),
+                            versionCode = it.packageMetadata.versionCode.toULong(),
                             versionName = it.packageMetadata.versionName,
                             buildApksResult = it.packageMetadata.buildApksResult.toByteArray(),
                             apks = it.packageMetadata.apkObjectMetadataMap
@@ -126,7 +126,7 @@ class AppEditPublicationRequestedProcessor(
                                     packageMetadataEntry {
                                         releaseChannel = channel.name.toReleaseChannel()
                                         packageMetadata = packageMetadata {
-                                            versionCode = channel.versionCode.toInt()
+                                            versionCode = channel.versionCode.toLong()
                                             versionName = channel.versionName
                                             buildApksResult =
                                                 BuildApksResult.parseFrom(channel.buildApksResult)
